[ADD] Perfumes Shop BE and EJS

+++ src/brands/brands.service.ts
@@ -1,27 +1,53 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 import { Brand } from './schemas/brand.schema';
 
 @Injectable()
 export class BrandsService {
-  constructor(@InjectModel(Brand.name) private brandModel: Model<Brand>) {}
+  constructor(
+    @InjectModel(Brand.name)
+    private readonly brandModel: Model<Brand>,
+  ) {}
 
-  async createBrand(brandName: string) {
-    return await new this.brandModel({ brandName }).save();
+  async createBrand(createData: any) {
+    return this.brandModel.create(createData);
   }
 
   async getAllBrands() {
-    return await this.brandModel.find();
+    return this.brandModel.find().sort({ createdAt: -1 }).lean();
   }
+  
 
-  async updateBrand(id: string, brandName: string) {
-    const brand = await this.brandModel.findByIdAndUpdate(id, { brandName }, { new: true });
+  async getBrandById(id: string) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new NotFoundException('Invalid Brand ID');
+    }
+    const brand = await this.brandModel.findById(id).lean();
     if (!brand) throw new NotFoundException('Brand not found');
     return brand;
   }
 
-  async deleteBrand(id: string) {
-    return await this.brandModel.findByIdAndDelete(id);
+  async getBrandByName(brandName: string): Promise<Brand | null> {
+    return this.brandModel.findOne({ brandName }).exec();
   }
+
+  async updateBrand(id: string, updateData: any) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new NotFoundException('Invalid Brand ID');
+    }
+    const brand = await this.brandModel.findByIdAndUpdate(id, updateData, { new: true }).lean();
+    if (!brand) throw new NotFoundException('Brand not found');
+    return brand;
+  }
+  
+  async deleteBrand(id: string) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new NotFoundException('Invalid Brand ID');
+    }
+    const result = await this.brandModel.findByIdAndDelete(id);
+    if (!result) throw new NotFoundException('Brand not found');
+    return { message: 'Brand deleted successfully' };
+  }
+  
 }
